requirements.txt
streamlit
pdfplumber
python-dotenv
langchain
langchain-openai
openai
requests
pandas